tiny few bits of cleanuo
