+++ apps/dotcom/sync-worker/src/routes/tla/screenshotTestHelpers.ts
@@ -1,0 +1,91 @@
+import { vi } from 'vitest'
+import { Environment } from '../../types'
+
+// Shared fakes for the Browser Run thumbnail / OG image tests (sharedBoardScreenshotMcp,
+// ogImageQueue, getOgImage). These were copy-pasted across those files; keep them here so the
+// R2/browser/queue fakes and token helpers stay in one place.
+
+// In-memory stand-in for the THUMBNAILS R2 bucket. Exposes `store` so tests can inspect or seed
+// entries directly. Covers get/head/put/delete; entries carry the customMetadata and upload time
+// the routes read.
+export function makeFakeThumbnailsBucket() {
+	const store = new Map<
+		string,
+		{ body: ArrayBuffer; customMetadata?: Record<string, string>; uploaded: Date }
+	>()
+	return {
+		store,
+		async get(key: string) {
+			const value = store.get(key)
+			if (!value) return null
+			return {
+				customMetadata: value.customMetadata,
+				uploaded: value.uploaded,
+				arrayBuffer: async () => value.body,
+			}
+		},
+		async head(key: string) {
+			const value = store.get(key)
+			if (!value) return null
+			return { customMetadata: value.customMetadata, uploaded: value.uploaded }
+		},
+		async put(
+			key: string,
+			body: ArrayBuffer,
+			options?: { customMetadata?: Record<string, string> }
+		) {
+			store.set(key, {
+				body,
+				customMetadata: options?.customMetadata,
+				uploaded: new Date(Date.now()),
+			})
+		},
+		async delete(key: string) {
+			store.delete(key)
+		},
+	}
+}
+
+// In-memory stand-in for the ROOMS bucket. Only `head` is used (to read the persisted snapshot's
+// etag); pass `null` to simulate a room with no persisted snapshot.
+export function makeFakeRoomsBucket(etag: string | null = 'room-etag-1') {
+	return {
+		async head(_key: string) {
+			return etag === null ? null : { etag }
+		},
+	}
+}
+
+// The BROWSER binding's `.quickAction('screenshot', body)` returns a Response whose body is the PNG
+// bytes. [1,2,3] base64-encodes to AQID. Pass a custom impl to simulate failures.
+export function makeBrowserBinding(
+	screenshot: (body: any) => Promise<Response> = async () =>
+		new Response(new Uint8Array([1, 2, 3]), { status: 200 })
+) {
+	return { quickAction: vi.fn((_action: string, body: any) => screenshot(body)) }
+}
+
+export function makeFakeQueue() {
+	return { send: vi.fn(async (_message: unknown) => undefined) }
+}
+
+export function makeScreenshotTestEnv(overrides: Partial<Record<string, unknown>> = {}) {
+	return {
+		BROWSER: makeBrowserBinding(),
+		MCP_SCREENSHOT_RENDER_ORIGIN: 'https://www.tldraw.com',
+		MCP_SCREENSHOT_TOKEN_SECRET: 'test-secret',
+		MEASURE: { writeDataPoint: vi.fn() },
+		QUEUE: makeFakeQueue(),
+		...overrides,
+	} as unknown as Environment
+}
+
+export function screenshotOf(env: Environment) {
+	return (env.BROWSER as any).quickAction as ReturnType<typeof vi.fn>
+}
+
+// quickAction is called as quickAction('screenshot', body); the render URL rides in body (arg 1).
+export function tokenFromScreenshot(env: Environment): string {
+	const body = screenshotOf(env).mock.calls[0]![1] as { url: string }
+	return new URL(body.url).searchParams.get('token')!
+}
